디버깅 완료, 인티저 처리못하는 문제 남음

+++ temp_libft.c
@@ -1,6 +1,8 @@
+# include <stdlib.h>
+
 void	free_twoarr(char **arr)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (arr[i])
